Remove more deprecations for 3.0
* Remove references to deprecated compaction strategies that don't exist
* Relocate unfortunately named `ConfigurableCompactionStrategy`, which
  existed solely to support the shell for user-initiated compactions,
  into two separate classes in accumulo-core.jar, as
  `ShellCompactCommandConfigurer` and `ShellCompactCommandSelector`
  1. Split into two classes to make it clear how they map to the SPI
     interfaces and to avoid confusing overloaded init methods with
     two different InitParameters classes that have the same simple name
  2. Put in accumulo-core.jar so it's on the classpath for servers
     that do compaction and the shell, so we can get the class name
     without using a String literal, which is fragile
  3. Avoid use of the "CompactionStrategy" naming, because that is the
     name of the deprecated, and now removed, interface
  4. Update corresponding tests
* Remove check for deprecated accumulo-site.xml, which was replaced in
  2.1. 3.x requires upgrade from at least 2.1, so that either will have
  already been updated in 2.1, or it will never have existed in 3.x.
  Also remove related config converter utilities.
* Remove randomizeVolumes from the admin command, which was previously
  made a noop and deprecated
* Remove unused accumulo-site.xml test resource
* Remove deprecated FateCommand from the shell
* Remove noop debug option from shell command-line that did nothing
  because its functionality has been replaced by the ability to
  dynamically configure log4j2 logging in 2.x
* Remove deprecated gc.trash.ignore (`GC_TRASH_IGNORE`) property, as
  follow-on to #3436; this includes removing a test that's no longer
  possible to succeed, since it's no longer possible to configure
  Accumulo to ignore Hadoop's trash settings; some comments were cleaned
  up in the remaining tests
