+++ core/src/test/java/org/apache/accumulo/core/compaction/ShellCompactCommandSelectorTest.java
@@ -1,0 +1,31 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.accumulo.core.compaction;
+
+import java.net.URISyntaxException;
+
+import org.junit.jupiter.api.Test;
+
+public class ShellCompactCommandSelectorTest {
+
+  @Test
+  public void testSelection() throws URISyntaxException {
+    // file selection options are adequately tested by ShellServerIT, so this is just a placeholder
+  }
+}
