--- server/manager/src/test/resources/conf/accumulo-site.xml
@@ -1,104 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      https://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
-
-<configuration>
-	<!--
-  Put your site-specific accumulo configurations here.
-
-  The available configuration values along with their defaults
-  are documented in docs/config.html
-
-  Unless you are simply testing at your workstation, you will most 
-  definitely need to change the three entries below.
-	-->
-
-  <property>
-    <name>instance.zookeeper.host</name>
-    <value>localhost:2181</value>
-    <description>comma separated list of zookeeper servers</description>
-  </property>
-
-  <property>
-    <name>instance.secret</name>
-    <value>DEFAULT</value>
-    <description>A secret unique to a given instance that all servers must know in order to
-      communicate with one another. Change it before initialization. To change it later use
-      ./bin/accumulo org.apache.accumulo.server.util.ChangeSecret, and then update this file.
-    </description>
-  </property>
-
-  <property>
-    <name>tserver.memory.maps.max</name>
-    <value>80M</value>
-  </property>
-
-  <property>
-    <name>tserver.cache.data.size</name>
-    <value>7M</value>
-  </property>
-
-  <property>
-    <name>tserver.cache.index.size</name>
-    <value>20M</value>
-  </property>
-
-  <property>
-    <name>tserver.sort.buffer.size</name>
-    <value>50M</value>
-  </property>
-
-  <property>
-    <name>tserver.wal.max.size</name>
-    <value>100M</value>
-  </property>
-
-  <property>
-    <name>general.classpaths</name>
-    <!--
-       Add the following for hadoop-2.0
-       $HADOOP_PREFIX/share/hadoop/common/.*.jar,
-       $HADOOP_PREFIX/share/hadoop/common/lib/.*.jar,
-       $HADOOP_PREFIX/share/hadoop/hdfs/.*.jar,
-       $HADOOP_PREFIX/share/hadoop/mapreduce/.*.jar,
-       $HADOOP_PREFIX/share/hadoop/yarn/.*.jar,
-    -->
-    <value>
-      $ACCUMULO_HOME/server/target/classes/,
-      $ACCUMULO_HOME/lib/accumulo-server.jar,
-      $ACCUMULO_HOME/core/target/classes/,
-      $ACCUMULO_HOME/lib/accumulo-core.jar,
-      $ACCUMULO_HOME/start/target/classes/,
-      $ACCUMULO_HOME/lib/accumulo-start.jar,
-      $ACCUMULO_HOME/proxy/target/classes/,
-      $ACCUMULO_HOME/lib/accumulo-proxy.jar,
-      $ACCUMULO_HOME/lib/[^.].*.jar,
-      $ZOOKEEPER_HOME/zookeeper[^.].*.jar,
-      $HADOOP_CONF_DIR,
-      $HADOOP_PREFIX/[^.].*.jar,
-      $HADOOP_PREFIX/lib/[^.].*.jar,
-    </value>
-    <description>Classpaths that accumulo checks for updates and class files.
-      When using the Security Manager, please remove the ".../target/classes/" values.
-    </description>
-  </property>
-</configuration>
